+++ basic.go
@@ -180,10 +180,17 @@ func LoopDiveSliceAndPtr(vValue reflect.Value) reflect.Value {
 
 // to check value is zero
 func IsZero(v reflect.Value) bool {
-	if v.Kind() == reflect.Struct && v.Type().Comparable() == false {
-		return reflect.Zero(v.Type()).String() == v.String()
-	} else {
-		return v.Interface() == reflect.Zero(v.Type()).Interface()
+	switch v.Kind() {
+	case reflect.Slice:
+		return v.IsNil()
+	case reflect.Map:
+		return v.IsNil()
+	default:
+		if v.Type().Comparable() == false {
+			return reflect.Zero(v.Type()).String() == v.String()
+		} else {
+			return v.Interface() == reflect.Zero(v.Type()).Interface()
+		}
 	}
 }