+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- [FIX] invoke IsZero with not struct value will panic
 - [NEW] add default tag match default value in database
 - [CHANGE] brick.Template method , the insertion part has changed